image-recognition/requirements.txt
torch
torchvision
fastai
Pillow
matplotlib
pandas
scikit-learn
jupyter
ipywidgets
